% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{Bacteria}
\alias{Bacteria}
\title{Bacteria}
\usage{
Bacteria(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, a new folder named Bacteria will be created and your working directory will be set into this folder, depending of the selected Kingdom.}
}
\value{
A new folder named Bacteria will be created and your working directory will be set into this folder, depending of the selected Kingdom.
}
\description{
This function create a new folder named Bacteria and set your working directory into this folder. Please, run setwd("Bacteria") after this function.
}
\examples{
\dontrun{Bacteria()}
# Please, run setwd("Bacteria") after this function.
}
