% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{clusteringGOs}
\alias{clusteringGOs}
\title{clusteringGOs}
\usage{
clusteringGOs(gen2go, div, cutHeight)
}
\arguments{
\item{gen2go}{from DESeq2 analysis pipeline}

\item{div}{div}

\item{cutHeight}{cutHeight}
}
\value{
a clustering GO
}
\description{
clusteringGOs from DESeq2 analysis pipeline
}
\examples{
\dontrun{clusteringGOs()}
}
