% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{dasva_raw_input}
\alias{dasva_raw_input}
\title{dasva_raw_input}
\usage{
dasva_raw_input(sampleTable, directory = ".", design, ignoreRank = FALSE, ...)
}
\arguments{
\item{sampleTable}{Depending of the heatmap_samples_hclust(), heatmap_samples_matrix(), PCA_data_dasva() and get_dasva() functions.}

\item{directory}{directory}

\item{design}{design}

\item{ignoreRank}{ignoreRank}

\item{...}{...}
}
\value{
object
}
\description{
Used in heatmap_samples_hclust(), heatmap_samples_matrix(), PCA_data_dasva() and get_dasva() functions.
}
\examples{
  \dontrun{dasva_raw <- dasva_raw_input(sampleTable = sampleTable,
  directory = targeted_analysis_dir,
  design= ~ condition)}
}
