% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{database_bacteria_creation}
\alias{database_bacteria_creation}
\title{database_bacteria_creation}
\usage{
database_bacteria_creation(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, the database will create itself.}
}
\value{
A data frame file named database_bacteria_package_all.tab created from the taxonomy_all_bacteria_QIIME2_and_NCBI_format.txt file and your own taxonomy_RepSeq.tsv file.
database_bacteria_creation()
}
\description{
Create a Database for Bacteria kingdom for Global analysis by Taxon_MWU analysis from targeted analysis. Please, run setwd("02_Global_analysis") after this function.
}
\examples{
# It is important not to write anything between the brackets, the database will create itself.
\dontrun{database_bacteria_creation()}
# Please, run setwd("02_Global_analysis") after this function.
}
