% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{database_fungi_creation}
\alias{database_fungi_creation}
\title{database_fungi_creation}
\usage{
database_fungi_creation(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, the database will create itself.}
}
\value{
A data frame file named database_fungi_package_all.tab created from the taxonomy_all_bacteria_QIIME2_and_NCBI_format.txt file and your own taxonomy.tsv file.
}
\description{
Create a Database for Fungi kingdom for Global analysis by Taxon_MWU analysis from targeted analysis only from rarefied ASVs. Please, run setwd("02_Global_analysis") after this function.
}
\examples{
# It is important not to write anything between the brackets, the database will create itself.
\dontrun{database_fungi_creation()}
# Please, run setwd("02_Global_analysis") after this function.
}
