% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{fisherTest}
\alias{fisherTest}
\title{fisherTest}
\usage{
fisherTest(gotable)
}
\arguments{
\item{gotable}{from gomwuStats from RBGOA}
}
\value{
fisherTest
}
\description{
Ficher Test from RBGOA
}
\examples{
\dontrun{fisherTest()}
}
