% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{format_input}
\alias{format_input}
\title{format_input}
\usage{
format_input(x)
}
\arguments{
\item{x}{Object from the Differential ASV abundance (DASVA) analysis}
}
\value{
an input for the input_global_analysis() function
}
\description{
Apply logP on both positive and negative ASVs FC
}
\examples{
\dontrun{format_input(x)}
}
