% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{Fungi}
\alias{Fungi}
\title{Fungi}
\usage{
Fungi(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, a new folder named Fungi will be created and your working directory will be set into this folder, depending of the selected Kingdom.}
}
\value{
A new folder named Fungi will be created and your working directory will be set into this folder, depending of the selected Kingdom.
}
\description{
This function create a new folder named Fungi and set your working directory into this folder. Please, run setwd("Fungi") after this function.
}
\examples{
\dontrun{Fungi()}
# plaese, run setwd("Fungi") after this function.
}
