% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{funguild_input_targeted}
\alias{funguild_input_targeted}
\title{funguild_input_targeted}
\usage{
funguild_input_targeted(x)
}
\arguments{
\item{x}{Object from the Differential ASV abundance (DASVA) analysis}
}
\value{
An Object used for Fungi Guilds informations for Fungi kingdom for targeted analysis from the Differential ASV abundance (DASVA) analysis
}
\description{
Prepare Object for Fungi Guilds for Fungi kingdom for targeted analysis
}
\examples{
\dontrun{get_funguilds_targeted(res_forest_vs_long_fallow_guilds)}
}
