% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_bactotraits_targeted}
\alias{get_bactotraits_targeted}
\title{get_bactotraits_targeted}
\usage{
get_bactotraits_targeted(x)
}
\arguments{
\item{x}{Object from the Differential ASV abundance (DASVA) analysis}
}
\value{
A data frame file with Bacterial Traits informations for Bacteria kingdom for targeted analysis from the Differential ASV abundance (DASVA) analysis
}
\description{
Obtain Bacterial Traits for Bacteria kingdom for targeted analysis
}
\examples{
\dontrun{get_bactotraits_targeted(res_forest_vs_long_fallow)}
}
