% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_dasva}
\alias{get_dasva}
\title{get_dasva}
\usage{
get_dasva(fitType = "")
}
\arguments{
\item{fitType}{Fit a Gamma-Poisson Generalized Linear Model, dispersion estimates for Negative Binomial distributed data, "parametric", local" or "mean"}
}
\value{
DASVA object
}
\description{
Creates the DASVA object. Fit a Gamma-Poisson Generalized Linear Model, dispersion estimates for Negative Binomial distributed data, "parametric", local" or "mean"
}
\examples{
\dontrun{dasva <- get_dasva(fitType="parametric")
dasva <- get_dasva(fitType="local")
dasva <- get_dasva(fitType="mean")}
}
