% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_funguilds}
\alias{get_funguilds}
\title{get_funguilds}
\usage{
get_funguilds(taxon_list_drawer)
}
\arguments{
\item{taxon_list_drawer}{object from get_taxon_list_drawer() function}
}
\value{
funguilds Object
}
\description{
get Fungi Guilds from taxon_list_drawer Object
}
\examples{
\dontrun{funguilds <- get_funguilds(taxon_list_drawer)}
}
