% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_funguilds_targeted}
\alias{get_funguilds_targeted}
\title{get_funguilds_targeted}
\usage{
get_funguilds_targeted(x)
}
\arguments{
\item{x}{Object from the funguild_input_targeted() output.}
}
\value{
A data frame file with Fungi Guilds informations for Fungi kingdom for targeted analysis from the Differential ASV abundance (DASVA) analysis
}
\description{
Obtain Fungi Guilds for Fungi kingdom for targeted analysis
}
\examples{
\dontrun{get_funguilds_targeted(res_forest_vs_long_fallow_guilds)}
}
