% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_input_files}
\alias{get_input_files}
\title{get_input_files}
\usage{
get_input_files(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
taxo
}
\description{
Created sub directory "Targeted_analysis" if not already exist. Then, create one file by condition into it, and then upload the taxonomy file. Please, run setwd("01_Targeted_analysis") after this function.
}
\examples{
\dontrun{taxo <- get_input_files()}
# please, run setwd("01_Targeted_analysis") after this function.
}
