% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_link_guilds}
\alias{get_link_guilds}
\title{get_link_guilds}
\usage{
get_link_guilds(taxon_list, funguilds)
}
\arguments{
\item{taxon_list}{object from taxon_mwu_list() function}

\item{funguilds}{object from get_funguilds() function}
}
\value{
link_guilds Object
}
\description{
get link guilds from taxon_list and funguilds Objects
}
\examples{
\dontrun{link_guilds <- get_link_guilds(taxon_list, funguilds)}
}
