% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{get_taxon_list_drawer}
\alias{get_taxon_list_drawer}
\title{get_taxon_list_drawer}
\usage{
get_taxon_list_drawer(taxon_list)
}
\arguments{
\item{taxon_list}{object from taxon_mwu_list() function}
}
\value{
taxon_list_drawer Object and "taxon_list_drawer_input.txt" file
}
\description{
get taxonomic list drawer
}
\examples{
\dontrun{taxon_list_drawer <- get_taxon_list_drawer(taxon_list)}
}
