% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{heatmap_condition}
\alias{heatmap_condition}
\title{heatmap_condition}
\usage{
heatmap_condition(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
Fill directly the annotation_col variable of the pheatmap() function
}
\description{
For Clustering step. Fill directly the annotation_col variable of the pheatmap() function
}
