% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{heatmap_data_dasva}
\alias{heatmap_data_dasva}
\title{heatmap_data_dasva}
\usage{
heatmap_data_dasva(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
Create the log2.norm.counts object.
}
\description{
For Clustering step. Create the log2.norm.counts object.
}
