% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{heatmap_samples_hclust}
\alias{heatmap_samples_hclust}
\title{heatmap_samples_hclust}
\usage{
heatmap_samples_hclust(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
hclust object for the heatmap.2() function
}
\description{
Adapt hclust for heatmap sample to sample analysis
}
\examples{
\dontrun{hc <- heatmap_samples_hclust()}
}
