% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{heatmap_samples_matrix}
\alias{heatmap_samples_matrix}
\title{heatmap_samples_matrix}
\usage{
heatmap_samples_matrix(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
samples matrix object for the heatmap.2() function
}
\description{
Adapt samples matrix for heatmap sample to sample analysis
}
\examples{
\dontrun{mat <- heatmap_samples_matrix()}
}
