% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{heatmap_taxo}
\alias{heatmap_taxo}
\title{heatmap_taxo}
\usage{
heatmap_taxo(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
log2.norm.counts_taxo used fro adding taxonomy in the pheatmap plot, instead of ASVs codes
}
\description{
Adding taxonomy in the pheatmap plot, instead of ASVs codes
}
\examples{
\dontrun{log2.norm.counts_taxo <- heatmap_taxo()}
}
