% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{input_global_analysis}
\alias{input_global_analysis}
\title{input_global_analysis}
\usage{
input_global_analysis(x)
}
\arguments{
\item{x}{Object from the Differential ASV abundance (DASVA) analysis}
}
\value{
Input Object for Global analysis by Taxon_MWU analysis from targeted analysis (I)
}
\description{
Input files creation for each condition for Global analysis by Taxon_MWU analysis from targeted analysis (I)
}
\examples{
\dontrun{input_global_analysis(res_forest_vs_long_fallow)}
}
