% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{Metazoan}
\alias{Metazoan}
\title{Metazoan}
\usage{
Metazoan(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, a new folder named Metazoan will be created and your working directory will be set into this folder, depending of the selected Kingdom.}
}
\value{
A new folder named Metazoan will be created and your working directory will be set into this folder, depending of the selected Kingdom.
}
\description{
This function create a new folder named Metazoan and set your working directory into this folder. Please, run setwd("Metazoan") after this function.
}
\examples{
\dontrun{Metazoan()}
# plaese, run setwd("Metazoan") after this function.
}
