% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{move_files}
\alias{move_files}
\title{move_files}
\usage{
move_files(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, files will move in the good folders, depending of the selected Kingdom before.}
}
\value{
Move the file in the good folders.
}
\description{
Move the file in the good folders. Depending on the previous Kingdom selection (e.g., Fungi 'Fungi()', Bacteria 'Bacteria()', etc.)
}
\examples{
\dontrun{move_files()}
}
