% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{mwuTest}
\alias{mwuTest}
\title{mwuTest}
\usage{
mwuTest(gotable, Alternative)
}
\arguments{
\item{gotable}{from gomwuStats from RBGOA}

\item{Alternative}{from gomwuStats from RBGOA}
}
\value{
mwuTest
}
\description{
Mann-Whitney U Test from RBGOA
}
\examples{
\dontrun{mwuTest()}
}
