% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{PCA_data_dasva}
\alias{PCA_data_dasva}
\title{PCA_data_dasva}
\usage{
PCA_data_dasva(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, all inputs will be adapted automatically.}
}
\value{
data. The PCA (Pincipal Component Analysis) data.
}
\description{
Compute the PCA (Pincipal Component Analysis) data.
}
