% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{plotDispASVs}
\alias{plotDispASVs}
\title{plotDispASVs}
\usage{
plotDispASVs(
  object,
  ymin,
  CV = FALSE,
  genecol = "black",
  fitcol = "red",
  finalcol = "dodgerblue",
  legend = TRUE,
  xlab,
  ylab,
  log = "xy",
  cex = 0.45,
  ...
)
}
\arguments{
\item{object}{Corresponding to the DASVA (Differential ASV abundance) object}

\item{ymin}{ymin}

\item{CV}{CV}

\item{genecol}{genecol}

\item{fitcol}{fitcol}

\item{finalcol}{finalcol}

\item{legend}{legend}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{log}{log}

\item{cex}{cex}

\item{...}{...}
}
\value{
A plot of Dispersion ASV
}
\description{
Create a plot of Dispersion ASV
}
\examples{
\dontrun{plotDispASVs(dasva)}
}
