% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{plotMA.dasva}
\alias{plotMA.dasva}
\title{plotMA.dasva}
\usage{
plotMA.dasva(
  object,
  alpha,
  main = "",
  xlab = "mean of normalized counts",
  ylim,
  MLE = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Object from the Differential ASV abundance (DASVA) analysis}

\item{alpha}{alpha}

\item{main}{main}

\item{xlab}{xlab}

\item{ylim}{ylim}

\item{MLE}{MLE}

\item{verbose}{verbose}

\item{...}{...}
}
\value{
A MA plot
}
\description{
Custom MA plots for the Differential ASV abundance (DASVA) analysis. defining a new function to plot all ASVs and not only log2FoldChange > 2
}
\examples{
\dontrun{plotMA.dasva(rXXX, main="XXX", ylim=c(-20,20))}
}
