% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{plotPCA.san}
\alias{plotPCA.san}
\title{plotPCA.san}
\usage{
plotPCA.san(object, intgroup = "condition", ntop = 500, returnData = FALSE)
}
\arguments{
\item{object}{An object use for the PCA}

\item{intgroup}{intgroup}

\item{ntop}{ntop}

\item{returnData}{returnData}
}
\value{
A PCA
}
\description{
Custom plotPCA function to plot PC1 et PC3
}
\examples{
\dontrun{plotPCA.san(object)}
}
