% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{plotSparsityASV}
\alias{plotSparsityASV}
\title{plotSparsityASV}
\usage{
plotSparsityASV(x, normalized = TRUE, ...)
}
\arguments{
\item{x}{Corresponding to the DASVA (Differential ASV abundance) object}

\item{normalized}{normalized}

\item{...}{...}
}
\value{
A plot of Sparsity ASV
}
\description{
Create a plot of Sparsity ASV
}
\examples{
\dontrun{plotSparsityASV(dasva)}
}
