% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{samplesInfo}
\alias{samplesInfo}
\title{samplesInfo}
\usage{
samplesInfo(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, comparisons will create themselves.}
}
\value{
a data.frame with conditions information from your SampleSheet_comparison.txt file, with focus on samplesInfo.
}
\description{
Imports conditions information from your SampleSheet_comparison.txt file, with focus on samplesInfo.
}
\examples{
\dontrun{samplesInfo <- samplesInfo()}
}
