% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{target_file}
\alias{target_file}
\title{target_file}
\usage{
target_file(nothing)
}
\arguments{
\item{nothing}{It's important not to write anything between the brackets, comparisons will create themselves.}
}
\value{
a data.frame with conditions information from your SampleSheet_comparison.txt file
}
\description{
Imports conditions information from your SampleSheet_comparison.txt file, with focus on iput files.
}
\examples{
\dontrun{target_file <- target_file()}
}
