% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anaconda.functions.R
\name{taxon_mwuStats}
\alias{taxon_mwuStats}
\title{taxon_mwuStats}
\usage{
taxon_mwuStats(
  input,
  goDatabase,
  goAnnotations,
  goDivision,
  Alternative = "t",
  adjust.multcomp = "BH",
  clusterCutHeight = 0.25,
  largest = 0.1,
  smallest = 5,
  perlPath = "perl",
  verbose = TRUE
)
}
\arguments{
\item{input}{input}

\item{goDatabase}{goDatabase}

\item{goAnnotations}{goAnnotations}

\item{goDivision}{goDivision}

\item{Alternative}{Alternative}

\item{adjust.multcomp}{adjust.multcomp}

\item{clusterCutHeight}{clusterCutHeight}

\item{largest}{largest}

\item{smallest}{smallest}

\item{perlPath}{perlPath}

\item{verbose}{verbose}
}
\value{
Statistical analysis for taxonomic rank
}
\description{
mwuStats from RBGOA adapted for taxonomic analysis
}
\examples{
\dontrun{taxon_mwuStats(input, ...)}
}
